change bigint to SERIAL in the sql file

+++ 01-init-db.sql
@@ -1,7 +1,7 @@
 -- CREATE TABLE Product
 DROP TABLE IF EXISTS product;
 CREATE TABLE product (
-                id bigint PRIMARY KEY,
+                id SERIAL PRIMARY KEY,
                 name VARCHAR(255),
                 description VARCHAR(255),
                 price double precision,
@@ -11,7 +11,7 @@ CREATE TABLE product (
 -- CREATE TABLE Client
 DROP TABLE IF EXISTS client;
 CREATE TABLE client (
-                        id bigint PRIMARY KEY,
+                        id SERIAL PRIMARY KEY,
                         client_name VARCHAR(255) NOT NULL UNIQUE,
                         password VARCHAR(255),
                         role VARCHAR(255)
